Hexvi: the basic prototype done
Added docstrings and migrated error handling to anyhow. Still, many
naive structure decisions are present, and individual functions need to
be thoroughly revised. Also, the performance of pancurses seems to be
terrible. A future solution might be to migrate to pure ncurses, or a
custom mechanism, based on something like termion.